+++ example/aes_example.dart
@@ -1,13 +1,13 @@
 import 'package:crypto_dart/crypto_dart.dart';
 
-void main(List<String> args) {
-  // Encrypt
-  var ciphertext =
-      CryptoDart.AES.encrypt('my message', 'secret key 123').toString();
-
-  // Decrypt
-  var bytes = CryptoDart.AES.decrypt(ciphertext, 'secret key 123');
-  var originalText = bytes.convertToString(CryptoDart.enc.Utf8);
-
-  print(originalText);
+void main() {
+  // Encrypt and decrypt data using AES
+  final key = 'ThisIsASecretKey';
+  final plainText = 'Hello, World!';
+  final encryptedText = CryptoDart.AES.encrypt(plainText, key);
+  final decryptedText = CryptoDart.AES
+      .decrypt(encryptedText, key)
+      .convertToString(CryptoDart.enc.Utf8);
+  print('Encrypted Text: $encryptedText');
+  print('Decrypted Text: $decryptedText');
 }