+++ lib/src/block_ciphers/tripledes.dart
@@ -112,7 +112,9 @@ class TripleDES extends cipher.BlockCipher {
     }
 
     if (key is String && options?.keyEncoding == null) {
-      var ctBytes = enc.Base64.parse(ciphertext as String);
+      var ctBytes = ciphertext is CipherParams
+          ? ciphertext.cipherText
+          : enc.Base64.parse(ciphertext as String);
       final cipherTextBytes =
           saltbytes == null ? ctBytes.sublist(_IV_SIZE) : ctBytes;
       saltbytes ??= ctBytes.sublist(_SALT_SIZE, _IV_SIZE);